+++ states_rebuilder_package/lib/src/when_connection_state.dart
@@ -31,10 +31,43 @@ class WhenRebuilder<T> extends StatelessWidget {
   ///have no error, and are not in the idle state, this callback will be invoked.
   final Widget Function(T data) onData;
 
-  //List of reactiveModels to observe
+  ///List of observable classes to which you want this [WhenRebuilder] to subscribe.
+  ///```dart
+  ///WhenRebuilder(
+  ///  models:[myModel1, myModel2, myModel3],
+  ///  onIdle: ()=> ...
+  ///  onWaiting: ()=> ...
+  ///  onError: (error)=> ...
+  ///  onData: (data)=> ...
+  ///)
+  ///```
+  ///
+  ///For the sake of performance consider using [observe] or [observeMany] instead.
   final List<ReactiveModel> models;
-  final StatesRebuilder Function() observe;
-  final List<StatesRebuilder Function()> observeMany;
+
+  ///an observable class to which you want [WhenRebuilder] to subscribe.
+  ///```dart
+  ///WhenRebuilder(
+  ///  observe:()=> myModel1,
+  ///  onIdle: ()=> ...
+  ///  onWaiting: ()=> ...
+  ///  onError: (error)=> ...
+  ///  onData: (data)=> ...
+  ///)
+  ///```
+  final ReactiveModel Function() observe;
+
+  ///List of observable classes to which you want this [WhenRebuilder] to subscribe.
+  ///```dart
+  ///WhenRebuilder(
+  ///  observeMany:[()=> myModel1,()=> myModel2,()=> myModel3],
+  ///  onIdle: ()=> ...
+  ///  onWaiting: ()=> ...
+  ///  onError: (error)=> ...
+  ///  onData: (data)=> ...
+  ///)
+  ///```
+  final List<ReactiveModel Function()> observeMany;
 
   ///A tag or list of tags you want this [WhenRebuilder] to register with.
   ///
@@ -47,8 +80,40 @@ class WhenRebuilder<T> extends StatelessWidget {
   ///Each [WhenRebuilder] has a default tag which is its [BuildContext]
   final dynamic tag;
 
+  ///ReactiveModel key used to control this widget from outside its [builder] method.
+  final RMKey rmKey;
+
+  ///```dart
+  ///WhenRebuilder(
+  ///  initState:(BuildContext context, ReactiveModel model)=> myModel.init([context,model]),
+  ///  observe:()=> myModel1,
+  ///  onIdle: ()=> ...
+  ///  onWaiting: ()=> ...
+  ///  onError: (error)=> ...
+  ///  onData: (data)=> ...
+  ///)
+  ///```
+  ///Called when this object is inserted into the tree.
   final void Function(BuildContext, ReactiveModel<T>) initState;
+
+  ///```dart
+  ///StateBuilder(
+  ///  dispose:(BuildContext context, ReactiveModel model) {
+  ///     myModel.dispose([context, model]);
+  ///   },
+  ///  observe:()=> myModel1,
+  ///  onIdle: ()=> ...
+  ///  onWaiting: ()=> ...
+  ///  onError: (error)=> ...
+  ///  onData: (data)=> ...
+  ///)
+  ///```
+  ///Called when this object is removed from the tree permanently.
   final void Function(BuildContext, ReactiveModel<T>) dispose;
+
+  ///Called whenever this widget is notified.
+  final dynamic Function(BuildContext context, ReactiveModel<T> model)
+      onSetState;
 
   const WhenRebuilder({
     Key key,
@@ -60,8 +125,10 @@ class WhenRebuilder<T> extends StatelessWidget {
     this.observe,
     this.observeMany,
     this.tag,
+    this.rmKey,
     this.initState,
     this.dispose,
+    this.onSetState,
   })  : assert(onWaiting != null),
         assert(onError != null),
         assert(onData != null),
@@ -74,8 +141,10 @@ class WhenRebuilder<T> extends StatelessWidget {
       observe: observe,
       observeMany: observeMany,
       tag: tag,
+      rmKey: rmKey,
       initState: initState,
       dispose: dispose,
+      onSetState: onSetState,
       builder: (context, modelRM) {
         bool isIdle = false;
         bool isWaiting = false;