+++ states_rebuilder_package/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.15.0 (2020-04-12)
-* Add `Inject.previous`. #47 [see more details](changelog/v-1.15.2.md) 
-* Addd Shortcuts to get and create model, future, stream `ReactiveModel`s: 
+* Add `Inject.previous`. #47 [see more details](changelog/v-1.15.0.md) 
+* Add Shortcuts to get and create model, future, stream `ReactiveModel`s: 
 ```dart
 IN.get<T>()//To consume the pure registered instance;
 RM.get<T>(); //to get the ReactiveModel instance of type T.
@@ -8,7 +8,9 @@ RM.create<T>(myModel);// to create a local ReactiveModel.
 RM.future<T>(myFuture);// to create a local future ReactiveModel.
 RM.stream<T>(myStream);// to create a local stream ReactiveModel.
 ```
-[see more details](changelog/v-1.15.2.md)
+[see more details](changelog/v-1.15.0.md)
+* Add `RM.getFuture` and `RM.getStream`, to get and create future or stream ReactiveModel.[see more details](changelog/v-1.15.0.md).
+* Add the concept of ReactiveModel key (RMKey). [see more details](changelog/v-1.15.0.md)
 * Fix issue #72
 
 ## 1.14.3 (2020-03-10)